+++ src/data/v2/stream.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2021 The apca Developers
+// Copyright (C) 2021-2022 The apca Developers
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use std::borrow::Borrow as _;
@@ -78,6 +78,9 @@ mod private {
 
 
 /// A trait representing the source from which to stream real time data.
+// TODO: Once we can use enumerations as const generic parameters we
+//       should probably switch over to repurposing `data::v2::Feed`
+//       here instead.
 pub trait Source: private::Sealed {
   /// Return a textual representation of the source.
   #[doc(hidden)]